update the test method to JUnit 5

+++ src/test/java/jock/demo/ProductsServiceTest.java
@@ -5,9 +5,8 @@ import jock.demo.model.Products;
 import jock.demo.service.BusinessException;
 import jock.demo.service.ProductsService;
 import jock.demo.service.ValidationException;
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
@@ -15,7 +14,6 @@ import javax.annotation.Resource;
 import java.util.Date;
 import java.util.List;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest
 public class ProductsServiceTest {
     @Resource
@@ -43,19 +41,19 @@ public class ProductsServiceTest {
 
         //insert a new product and if add success the products should have id.
         Products products = productsService.addComingSoonProducts(p1);
-        Assert.assertNotEquals(products.getId().longValue(), 0L);
+        Assertions.assertNotEquals(products.getId().longValue(), 0L);
 
         //find the new list size
         List<Products> list1 = productsService.getComingSoonProduct();
         Integer newSize = list1.size();
         // it should be greater than origin size
-        Assert.assertNotEquals(originSize, newSize);
+        Assertions.assertNotEquals(originSize, newSize);
 
         //find by product name
         p2 = productsMapper.selectByPrimaryName("test1");
         //it should be not null
-        Assert.assertNotNull(p2);
-        Assert.assertEquals(p2.getName(),"test1");
+        Assertions.assertNotNull(p2);
+        Assertions.assertEquals(p2.getName(),"test1");
 
         //remove the product
         productsService.removeComingSoonProducts(p2.getId());
@@ -64,7 +62,7 @@ public class ProductsServiceTest {
         //the list size should be equals to the origin size
         list1 = productsService.getComingSoonProduct();
         newSize = list1.size();
-        Assert.assertEquals(originSize, newSize);
+        Assertions.assertEquals(originSize, newSize);
     }
 
     /**
@@ -77,7 +75,10 @@ public class ProductsServiceTest {
         p1.setDescription("just a test");
         //clear
         Products p2 = productsMapper.selectByPrimaryName("test2");
-        productsService.removeCurrentProducts(p2.getId());
+        if(p2!=null){
+            productsService.removeCurrentProducts(p2.getId());
+
+        }
 
 
         //record origin of list size
@@ -86,7 +87,7 @@ public class ProductsServiceTest {
 
         //insert a new product and if add success the products should have id.
         Products products = productsService.addCurrentProducts(p1);
-        Assert.assertNotEquals(products.getId().longValue(), 0L);
+        Assertions.assertNotEquals(products.getId().longValue(), 0L);
 
 
         //find the new list size
@@ -94,14 +95,14 @@ public class ProductsServiceTest {
         Integer newSize = list1.size();
 
         // it should be greater than origin size
-        Assert.assertNotEquals(originSize, newSize);
-        Assert.assertNotNull(list1);
+        Assertions.assertNotEquals(originSize, newSize);
+        Assertions.assertNotNull(list1);
 
         //find by product name
         p2 = productsMapper.selectByPrimaryName("test2");
         //it should be not null
-        Assert.assertNotNull(p2);
-        Assert.assertEquals(p2.getName(),"test2");
+        Assertions.assertNotNull(p2);
+        Assertions.assertEquals(p2.getName(),"test2");
 
         //remove the product
         productsService.removeCurrentProducts(p2.getId());
@@ -109,26 +110,26 @@ public class ProductsServiceTest {
         //the list size should be equals to the origin size
         list1 = productsService.getAllCurrentProducts();
         newSize = list1.size();
-        Assert.assertEquals(originSize, newSize);
+        Assertions.assertEquals(originSize, newSize);
 
     }
 
-    @Test(expected = ValidationException.class)
+    @Test
     public void currentAddNullTest() {
-        productsService.addCurrentProducts(null);
+        Assertions.assertThrows(ValidationException.class,()->productsService.addCurrentProducts(null));
     }
 
     /**
      * the same data add twice will occur exception
      */
-    @Test(expected = BusinessException.class)
+    @Test
     public void currentAddTwiceTest() {
         Products p1 = new Products();
         p1.setName("test2");
         p1.setDescription("just a test");
         productsService.addCurrentProducts(p1);
-        productsService.addCurrentProducts(p1);
 
+        Assertions.assertThrows(BusinessException.class,()->productsService.addCurrentProducts(p1));
         //clear
         Products p2 = productsMapper.selectByPrimaryName("test2");
         productsService.removeCurrentProducts(p2.getId());
@@ -137,7 +138,7 @@ public class ProductsServiceTest {
     /**
      * the same data remove twice will occur exception
      */
-    @Test(expected = BusinessException.class)
+    @Test
     public void currentRemoveTwiceTest() {
         Products p1 = new Products();
         p1.setName("test2");
@@ -145,40 +146,38 @@ public class ProductsServiceTest {
         p1 = productsService.addCurrentProducts(p1);
 
         productsService.removeCurrentProducts(p1.getId());
-
-        productsService.removeCurrentProducts(p1.getId());
-
+        Products finalP = p1;
+        Assertions.assertThrows(BusinessException.class,()->productsService.removeCurrentProducts(finalP.getId()));
 
     }
 
-    @Test(expected = ValidationException.class)
+    @Test
     public void currentAddEmptyTest() {
         Products products = new Products();
-        productsService.addCurrentProducts(products);
+        Assertions.assertThrows(ValidationException.class,()->productsService.addCurrentProducts(products));
     }
 
-    @Test(expected = ValidationException.class)
+    @Test
     public void soonAddNullTest() {
-        productsService.addComingSoonProducts(null);
+        Assertions.assertThrows(ValidationException.class,()->productsService.addComingSoonProducts(null));
     }
 
-    @Test(expected = ValidationException.class)
+    @Test
     public void soonAddEmptyTest() {
         Products products = new Products();
-        productsService.addComingSoonProducts(products);
+        Assertions.assertThrows(ValidationException.class,()->productsService.addComingSoonProducts(products));
     }
 
     /**
      * the same data add twice will occur exception
      */
-    @Test(expected = BusinessException.class)
+    @Test
     public void soonAddTwiceTest() {
         Products p1 = new Products();
         p1.setName("test1");
         p1.setDescription("just a test");
         productsService.addComingSoonProducts(p1);
-        productsService.addComingSoonProducts(p1);
-
+        Assertions.assertThrows(BusinessException.class,()->productsService.addComingSoonProducts(p1));
         //clear
         Products p2 = productsMapper.selectByPrimaryName("test1");
         productsService.removeCurrentProducts(p2.getId());
@@ -188,7 +187,7 @@ public class ProductsServiceTest {
     /**
      * the same data remove twice will occur exception
      */
-    @Test(expected = BusinessException.class)
+    @Test
     public void soonRemoveTwiceTest() {
         Products p1 = new Products();
         p1.setName("test1");
@@ -197,28 +196,27 @@ public class ProductsServiceTest {
 
         productsService.removeComingSoonProducts(p1.getId());
 
-        productsService.removeComingSoonProducts(p1.getId());
-
-
+        Products finalP = p1;
+        Assertions.assertThrows(BusinessException.class,()->productsService.removeComingSoonProducts(finalP.getId()));
     }
 
-    @Test(expected = ValidationException.class)
+    @Test
     public void currentRemoveNullTest() {
-        productsService.removeCurrentProducts(null);
+        Assertions.assertThrows(ValidationException.class,()->productsService.removeCurrentProducts(null));
     }
 
-    @Test(expected = BusinessException.class)
+    @Test
     public void currentRemoveEmptyTest() {
-        productsService.removeCurrentProducts(Integer.MAX_VALUE);
+        Assertions.assertThrows(BusinessException.class,()->productsService.removeCurrentProducts(Integer.MAX_VALUE));
     }
 
-    @Test(expected = ValidationException.class)
+    @Test
     public void soonRemoveNullTest() {
-        productsService.removeComingSoonProducts(null);
+        Assertions.assertThrows(ValidationException.class,()->productsService.removeComingSoonProducts(null));
     }
 
-    @Test(expected = BusinessException.class)
+    @Test
     public void soonRemoveEmptyTest() {
-        productsService.removeComingSoonProducts(Integer.MAX_VALUE);
+        Assertions.assertThrows(BusinessException.class,()->productsService.removeComingSoonProducts(Integer.MAX_VALUE));
     }
 }
